+++ lib/commands/crdt/fetchcounter.js
